Client/server: a lot of improvements
git-svn-id: http://talendforge.org/svn/tos/trunk@3341 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.commons/src/main/java/org/talend/commons/utils/threading/ExecutionLimiter.java
@@ -34,7 +34,7 @@ public abstract class ExecutionLimiter {
 
     private boolean inExecution;
 
-    private int timeBeforeNewExecution;
+    private long timeBeforeNewExecution;
 
     private long startTime;
 
@@ -46,11 +46,11 @@ public abstract class ExecutionLimiter {
         super();
     }
 
-    public ExecutionLimiter(int timeBeforeNewExecute) {
+    public ExecutionLimiter(long timeBeforeNewExecute) {
         this.timeBeforeNewExecution = timeBeforeNewExecute;
     }
 
-    public ExecutionLimiter(int timeBeforeNewExecute, boolean finalExecute) {
+    public ExecutionLimiter(long timeBeforeNewExecute, boolean finalExecute) {
         this.timeBeforeNewExecution = timeBeforeNewExecute;
         this.finalExecute = finalExecute;
     }
@@ -163,7 +163,7 @@ public abstract class ExecutionLimiter {
         return returnValue;
     }
 
-    public int getTimeBeforeNewExecution() {
+    public long getTimeBeforeNewExecution() {
         return timeBeforeNewExecution;
     }
